Extract stattest functions in separate folder for reuse.
