+++ evidently/analyzers/data_drift_analyzer.py
@@ -2,12 +2,12 @@
 # coding: utf-8
 import pandas as pd
 import numpy as np
-from scipy.stats import ks_2samp, chisquare
 
 from evidently import ColumnMapping
 from evidently.analyzers.base_analyzer import Analyzer
 from evidently.options import DataDriftOptions
-from .utils import proportions_diff_z_stat_ind, proportions_diff_z_test, process_columns
+from evidently.analyzers.stattests import chi_stat_test, ks_stat_test, z_stat_test
+from evidently.analyzers.utils import process_columns
 
 
 def dataset_drift_evaluation(p_values, confidence=0.95, drift_share=0.5):
@@ -34,8 +34,6 @@ class DataDriftAnalyzer(Analyzer):
         nbinsx = options.nbinsx
         confidence = options.confidence
         drift_share = options.drift_share
-
-        result['options'] = options.as_dict()
         # calculate result
         result['metrics'] = {}
 
@@ -45,9 +43,6 @@ class DataDriftAnalyzer(Analyzer):
             func = None if options.feature_stattest_func is None \
                 else options.feature_stattest_func.get(feature_name, None)
             func = options.stattest_func if func is None else func
-
-            def ks_stat_test(reference_data, current_data):
-                return ks_2samp(reference_data, current_data)[1]
             func = ks_stat_test if func is None else func
             p_value = func(reference_data[feature_name], current_data[feature_name])
             p_values.append(p_value)
@@ -105,33 +100,3 @@ class DataDriftAnalyzer(Analyzer):
         result['metrics']['share_drifted_features'] = share_drifted_features
         result['metrics']['dataset_drift'] = dataset_drift
         return result
-
-
-def chi_stat_test(reference_data, current_data):
-    ref_feature_vc = reference_data[np.isfinite(reference_data)].value_counts()
-    current_feature_vc = current_data[np.isfinite(current_data)].value_counts()
-    keys = set(list(reference_data[np.isfinite(reference_data)].unique()) +
-               list(current_data[np.isfinite(current_data)].unique()))
-
-    ref_feature_dict = dict.fromkeys(keys, 0)
-    for key, item in zip(ref_feature_vc.index, ref_feature_vc.values):
-        ref_feature_dict[key] = item
-
-    current_feature_dict = dict.fromkeys(keys, 0)
-    for key, item in zip(current_feature_vc.index, current_feature_vc.values):
-        current_feature_dict[key] = item
-    f_exp = [value[1] for value in sorted(ref_feature_dict.items())]
-    f_obs = [value[1] for value in sorted(current_feature_dict.items())]
-    return chisquare(f_exp, f_obs)[1]
-
-
-def z_stat_test(reference_data, current_data):
-    keys = set(list(reference_data[np.isfinite(reference_data)].unique()) +
-               list(current_data[np.isfinite(current_data)].unique()))
-    ordered_keys = sorted(list(keys))
-    return proportions_diff_z_test(
-        proportions_diff_z_stat_ind(
-            reference_data.apply(lambda x, key=ordered_keys[0]: 0 if x == key else 1),
-            current_data.apply(lambda x, key=ordered_keys[0]: 0 if x == key else 1)
-        )
-    )
