+++ evidently/analyzers/stattests/__init__.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python
+# coding: utf-8
+from .ks_stattest import ks_stat_test
+from .z_stattest import z_stat_test
+from .chisquare_stattest import chi_stat_test
