+++ evidently/analyzers/utils.py
@@ -2,35 +2,8 @@ from typing import Optional, Union, Sequence
 
 import numpy as np
 import pandas as pd
-from scipy.stats import norm
 
 from evidently.pipeline.column_mapping import ColumnMapping
-
-
-def proportions_diff_z_stat_ind(ref, curr):
-    # pylint: disable=invalid-name
-    n1 = len(ref)
-    n2 = len(curr)
-
-    p1 = float(sum(ref)) / n1
-    p2 = float(sum(curr)) / n2
-    P = float(p1 * n1 + p2 * n2) / (n1 + n2)
-
-    return (p1 - p2) / np.sqrt(P * (1 - P) * (1. / n1 + 1. / n2))
-
-
-def proportions_diff_z_test(z_stat, alternative='two-sided'):
-    if alternative == 'two-sided':
-        return 2 * (1 - norm.cdf(np.abs(z_stat)))
-
-    if alternative == 'less':
-        return norm.cdf(z_stat)
-
-    if alternative == 'greater':
-        return 1 - norm.cdf(z_stat)
-
-    raise ValueError("alternative not recognized\n"
-                     "should be 'two-sided', 'less' or 'greater'")
 
 
 class DatasetUtilityColumns:
