+++ evidently/options/data_drift.py
@@ -12,12 +12,3 @@ class DataDriftOptions:
     feature_stattest_func: Optional[Dict[str, Callable]] = None
     cat_target_stattest_func: Optional[Callable] = None
     num_target_stattest_func: Optional[Callable] = None
-
-
-    def as_dict(self):
-        return {
-            "confidence": self.confidence,
-            "drift_share": self.drift_share,
-            "nbinsx": self.nbinsx,
-            "xbins": self.xbins
-        }
